refactor(test-utils): extract createTableMock for multi-delete dialog tests
Add shared TanStack Table mock in test-utils/tanstack-table; use it from
tasks and users multi-delete dialog tests.
